report_to_otx/requirements.txt
OTXv2
PyYAML
